refactor go routes/ handler to make it easier to manage
